+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Yii Framework 2 redis extension Change Log
 
 - Bug #114: Fixed ActiveQuery `not between` and `not` conditions which where not working correctly (cebe, ak1987)
 - Bug #123: Fixed ActiveQuery to work with negative limit values, which are used in ActiveDataProvider for the count query (cebe)
+- Chg: Added PHP 7.2 compatibility (brandonkelly)
 
 
 2.0.6 April 05, 2017